+++ packages/eslint-config/src/config/plugins/javascript.ts
@@ -12,7 +12,7 @@ export default createConfig<OptionsPackageJson>("js", async (config) => {
     return [
         {
             languageOptions: {
-                ecmaVersion: 2022,
+                ecmaVersion: "latest",
                 globals: {
                     ...globals.browser,
                     ...globals.es2021,
@@ -38,7 +38,7 @@ export default createConfig<OptionsPackageJson>("js", async (config) => {
                     ecmaFeatures: {
                         jsx: true,
                     },
-                    ecmaVersion: 2022,
+                    ecmaVersion: "latest",
                     sourceType: "module",
                 },
                 sourceType: "module",