+++ packages/eslint-config/src/utils/interop-default.ts
@@ -3,7 +3,7 @@ import type { Awaitable } from "../types";
 const interopDefault = async <T>(m: Awaitable<T>): Promise<T extends { default: infer U } ? U : T> => {
     const resolved = await m;
 
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-return,@typescript-eslint/no-unsafe-member-access
     return (resolved as any).default ?? resolved;
 };