+++ packages/lint-staged-config/__tests__/index.test.ts
@@ -8,7 +8,6 @@ import {
 
 import { defineConfig, eslintExtensions } from "../src";
 
-// eslint-disable-next-line vitest/hoisted-apis-on-top
 const { existsSyncMock, findPackageManagerSyncMock, parsePackageJsonMock } = vi.hoisted(() => {
     return {
         // eslint-disable-next-line vitest/require-mock-type-parameters